cros-disks: Pass mount flags as uint64
The type int is too short to handle all the bits of the mount flags.

BUG=None
TEST=CQ

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3994664
Tested-by: François Degros <fdegros@chromium.org>
Commit-Queue: François Degros <fdegros@chromium.org>

+++ cros-disks/fuse_mounter.cc
@@ -266,7 +266,7 @@ std::unique_ptr<MountPoint> FUSEMounter::Mount(
   }
 
   // Prepare mount flags.
-  int mount_flags = MS_NODEV | MS_NOSUID | MS_NOEXEC | MS_DIRSYNC;
+  uint64_t mount_flags = MS_NODEV | MS_NOSUID | MS_NOEXEC | MS_DIRSYNC;
 
   if (read_only)
     mount_flags |= MS_RDONLY;
